+++ packages/datasource-customizer/src/decorators/sort-emulate/collection.ts
@@ -2,7 +2,6 @@ import {
   Caller,
   CollectionDecorator,
   CollectionSchema,
-  ColumnSchema,
   ConditionTreeFactory,
   DataSourceDecorator,
   FieldSchema,
@@ -20,22 +19,33 @@ import {
 export default class SortEmulate extends CollectionDecorator {
   override readonly dataSource: DataSourceDecorator<SortEmulate>;
   private readonly sorts: Map<string, Sort> = new Map();
+  private readonly disabledSorts: Set<string> = new Set();
 
   emulateFieldSorting(name: string): void {
-    this.replaceFieldSorting(name, null);
+    this.replaceOrEmulateFieldSorting(name, null);
   }
 
+  /**
+   * Disable sorting on this field. This only prevents the end-user to sort on this field.
+   * It will still be possible to sort on this field in the customizations code.
+   * @param name name of the field
+   * @deprecated this method will be removed soon.
+   */
   disableFieldSorting(name: string): void {
-    // This is not the best way to achieve the disable behavior
-    // but basically we replace sort with nothing to disable any sort
-    this.replaceFieldSorting(name, []);
+    FieldValidator.validate(this, name);
+
+    this.disabledSorts.add(name);
+    this.markSchemaAsDirty();
   }
 
   replaceFieldSorting(name: string, equivalentSort: PlainSortClause[]): void {
-    FieldValidator.validate(this, name);
+    if (!equivalentSort)
+      throw new Error('A new sorting method should be provided to replace field sorting');
+    this.replaceOrEmulateFieldSorting(name, equivalentSort);
+  }
 
-    const field = this.childCollection.schema.fields[name] as ColumnSchema;
-    if (!field) throw new Error('Cannot replace sort on relation');
+  private replaceOrEmulateFieldSorting(name: string, equivalentSort: PlainSortClause[]): void {
+    FieldValidator.validate(this, name);
 
     this.sorts.set(name, equivalentSort ? new Sort(...equivalentSort) : null);
     this.markSchemaAsDirty();
@@ -81,16 +91,21 @@ export default class SortEmulate extends CollectionDecorator {
     const fields: Record<string, FieldSchema> = {};
 
     for (const [name, schema] of Object.entries(childSchema.fields)) {
-      const fieldSort = this.sorts.get(name);
+      if (schema.type === 'Column') {
+        let sortable = schema.isSortable;
 
-      fields[name] =
-        this.sorts.has(name) &&
-        schema.type === 'Column' &&
-        // In order to support disableFieldSorting with empty array (isSortable: false)
-        // and emulateFieldSorting with null value (isSortable: true)
-        (!fieldSort || (fieldSort && fieldSort?.length !== 0))
-          ? { ...schema, isSortable: true }
-          : schema;
+        if (this.disabledSorts.has(name)) {
+          // disableFieldSorting
+          sortable = false;
+        } else if (this.sorts.has(name)) {
+          // replaceFieldSorting
+          sortable = true;
+        }
+
+        fields[name] = { ...schema, isSortable: sortable };
+      } else {
+        fields[name] = schema;
+      }
     }
 
     return { ...childSchema, fields };