+++ packages/_example/CHANGELOG.md
@@ -1,3 +1,61 @@
+## example [1.6.77](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.6.76...example@1.6.77) (2024-01-23)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/agent:** upgraded to 1.36.17
+
+## example [1.6.76](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.6.75...example@1.6.76) (2024-01-22)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/agent:** upgraded to 1.36.16
+* **@forestadmin/datasource-dummy:** upgraded to 1.0.94
+
+## example [1.6.75](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.6.74...example@1.6.75) (2024-01-22)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/agent:** upgraded to 1.36.15
+* **@forestadmin/datasource-dummy:** upgraded to 1.0.93
+
+## example [1.6.74](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.6.73...example@1.6.74) (2024-01-18)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/agent:** upgraded to 1.36.14
+* **@forestadmin/datasource-dummy:** upgraded to 1.0.92
+
+## example [1.6.73](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.6.72...example@1.6.73) (2024-01-17)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/agent:** upgraded to 1.36.13
+* **@forestadmin/datasource-dummy:** upgraded to 1.0.91
+* **@forestadmin/datasource-mongoose:** upgraded to 1.5.33
+* **@forestadmin/datasource-sequelize:** upgraded to 1.5.27
+* **@forestadmin/datasource-sql:** upgraded to 1.7.44
+* **@forestadmin/datasource-toolkit:** upgraded to 1.29.2
+
 ## example [1.6.72](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.6.71...example@1.6.72) (2024-01-03)