+++ packages/datasource-sequelize/CHANGELOG.md
@@ -1,3 +1,13 @@
+## @forestadmin/datasource-sequelize [1.5.27](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-sequelize@1.5.26...@forestadmin/datasource-sequelize@1.5.27) (2024-01-17)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-toolkit:** upgraded to 1.29.2
+
 ## @forestadmin/datasource-sequelize [1.5.26](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-sequelize@1.5.25...@forestadmin/datasource-sequelize@1.5.26) (2023-12-12)