+++ packages/plugin-aws-s3/CHANGELOG.md
@@ -1,3 +1,44 @@
+## @forestadmin/plugin-aws-s3 [1.3.58](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-aws-s3@1.3.57...@forestadmin/plugin-aws-s3@1.3.58) (2024-01-22)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.40.4
+
+## @forestadmin/plugin-aws-s3 [1.3.57](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-aws-s3@1.3.56...@forestadmin/plugin-aws-s3@1.3.57) (2024-01-22)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.40.3
+
+## @forestadmin/plugin-aws-s3 [1.3.56](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-aws-s3@1.3.55...@forestadmin/plugin-aws-s3@1.3.56) (2024-01-18)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.40.2
+
+## @forestadmin/plugin-aws-s3 [1.3.55](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-aws-s3@1.3.54...@forestadmin/plugin-aws-s3@1.3.55) (2024-01-17)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.40.1
+* **@forestadmin/datasource-toolkit:** upgraded to 1.29.2
+
 ## @forestadmin/plugin-aws-s3 [1.3.54](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-aws-s3@1.3.53...@forestadmin/plugin-aws-s3@1.3.54) (2024-01-03)